Added contributions by Tom Acree (LRS support).
The SDOGeometryType has been almost completely rewritten.
This commit is an intermediate step in the process of 
eliminating the dependency on sdoapi.jar. 

Unit test for spatial queries is now adapted to the new 
version of the delegate in Hibernate Spatial core.

The dialect creates for the spatial relation expressions 
SQL expressions that they work both in SQL select and where clauses.